+++ docs/architecture/adr-map.md
@@ -276,7 +276,7 @@ review`; evidence is the number of declared qualification references.
 | KF · ec85 | [Separate Primitive definition, admission, and availability planes](../adr/KF-ADR-019f9659-5560-7d9d-a143-78766f19ec85.md) | primitive-management-three-plane | accepted / implemented / self-reviewed | 5 | [KF · 8aec](../adr/KF-ADR-019f917f-d116-70e8-b4a1-2e0209598aec.md) (shared terms: management, plane, primitive)<br>[SHIFU · f011](../adr/SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011.md) (shared terms: plane)<br>[KF · fe03](../adr/KF-ADR-019f86da-4f90-7d81-90a0-d144fc27fe03.md) (shared terms: admission) |
 | KF · 0429 | [Portable format authority is a composition contract](../adr/KF-ADR-019f96a2-c686-76e1-9261-f6106aa50429.md) | portable-format-authority | accepted / staged / self-reviewed | 50 | [KF · d844](../adr/KF-ADR-019f86da-4f90-7008-b154-d488eb39d844.md) (shared terms: authority, composition)<br>[SHIFU · 7ace](../adr/SHIFU-ADR-019f86ff-a8d6-7431-ae05-0ec95fdb7ace.md) (shared terms: authority)<br>[KF · 7733](../adr/KF-ADR-019f8822-1d7a-7594-adea-65ad12c47733.md) (shared terms: contract)<br>[KF · 38a9](../adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md) (shared terms: authority) |
 | KF · b9bd | [Kungfu owns native Work Control vocabulary](../adr/KF-ADR-019f9771-4c20-7e2c-8e7c-3f3cb3f1b9bd.md) | kungfu-native-work-control | accepted / implemented / self-reviewed | 27 | [KF · 5d62](../adr/KF-ADR-019fee22-e71d-7da9-8a44-9403c21a5d62.md) (shared terms: kungfu, work)<br>[KF · d8bc](../adr/KF-ADR-019f86da-4f90-7a66-b427-f4bcd638d8bc.md) (shared terms: kungfu)<br>[KF · d412](../adr/KF-ADR-019f86da-4f90-7f7b-90be-c002b024d412.md) (shared terms: control)<br>[KF · cb76](../adr/KF-ADR-019f86da-4f90-713d-8626-d70bca82cb76.md) (shared terms: kungfu) |
-| KF · 7ec0 | [Maintainability governance uses read-only composite regression ratchets](../adr/KF-ADR-019f9f04-f8bd-7002-a702-1b9f66827ec0.md) | maintainability-governance | accepted / implemented / self-reviewed | 82 | [KF · a40c](../adr/KF-ADR-019f8fb8-579b-78d5-9ebb-da03bb9aa40c.md) (shared terms: governance) |
+| KF · 7ec0 | [Maintainability governance uses read-only composite regression ratchets](../adr/KF-ADR-019f9f04-f8bd-7002-a702-1b9f66827ec0.md) | maintainability-governance | accepted / implemented / self-reviewed | 86 | [KF · a40c](../adr/KF-ADR-019f8fb8-579b-78d5-9ebb-da03bb9aa40c.md) (shared terms: governance) |
 | KF · 90d2 | [First-party longitudinal evolution map](../adr/KF-ADR-019fa773-aae4-7f50-80a5-ce53b08490d2.md) | first-party-longitudinal-evolution-map | accepted / implemented / self-reviewed | 7 | [KF · 86f1](../adr/KF-ADR-019f86da-4f90-7e5e-ae22-2a8fc24086f1.md) (shared terms: first) |
 | KF · 38a9 | [Exit history surfaces are observers over unchanged domain authority](../adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md) | durable-history-data-protection | accepted / staged / self-reviewed | 9 | [KF · dccd](../adr/KF-ADR-019fbee3-68b1-7154-a6d2-860b9506dccd.md) (same theme: durable-history-data-protection; shared terms: data, domain, durable, history)<br>[KF · d844](../adr/KF-ADR-019f86da-4f90-7008-b154-d488eb39d844.md) (shared terms: authority, domain, exit)<br>[SHIFU · 7ace](../adr/SHIFU-ADR-019f86ff-a8d6-7431-ae05-0ec95fdb7ace.md) (shared terms: authority)<br>[KF · 7733](../adr/KF-ADR-019f8822-1d7a-7594-adea-65ad12c47733.md) (shared terms: domain) |
 | KF · dccd | [Durable history qualification is a bounded cross-domain proof](../adr/KF-ADR-019fbee3-68b1-7154-a6d2-860b9506dccd.md) | durable-history-data-protection | accepted / staged / self-reviewed | 4 | [KF · 38a9](../adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md) (same theme: durable-history-data-protection; shared terms: data, domain, durable, history)<br>[KF · 7733](../adr/KF-ADR-019f8822-1d7a-7594-adea-65ad12c47733.md) (shared terms: domain)<br>[KF · d844](../adr/KF-ADR-019f86da-4f90-7008-b154-d488eb39d844.md) (shared terms: domain)<br>[KF · bba2](../adr/KF-ADR-019f86da-4f90-7834-a788-5b65c6a4bba2.md) (shared terms: bounded) |
